add top-level `alert` property to template
When using the CLI, the alert will be shown to the user if the template is used and if the alert is non-empty. Printing is done to stderr to play nicely when piping output to files or build commands. Users are prompted for confirmation.

This is especially useful for non-free software, like FSL. Users can be notified of licenses or other particulars of the software.